Support ipykernel 6 series and switching to async
Starting with version 6 ipykernel switched to async operation making it hard to
support synchronous interface that worked well enough with the version 5 of
ipykernel.

I don't see an easy way to support both 5 an 6 ipykernel in the same library, so
the new version will only be compatible with 6 series of ipykernel.